chore: update charting endpoints + tests
