+++ src/accounts/account.v3.controller.spec.ts
@@ -1,7 +1,7 @@
 import { PlatformTest } from '@tsed/common';
 import SuperTest from 'supertest';
 
-import { NetworkStatus } from '../errors/enums/newtroks.status.enum';
+import { NetworkStatus } from '../errors/enums/network-status.enum';
 import { Server } from '../Server';
 import { TEST_ADDR } from '../test/constants';
 import { mockBadgerSdk, setupMockAccounts } from '../test/tests.utils';
