--- src/charts/enums/chart-granularity.enum.ts
@@ -1,4 +1,0 @@
-export enum ChartGranularity {
-  HOUR = 'hour',
-  DAY = 'day',
-}
