+++ src/errors/allocation/not.found.error.ts
@@ -1,6 +1,6 @@
 import { BaseApiError } from '../base.error';
 import { ApiErrorCode } from '../enums/error.codes.enum';
-import { NetworkStatus } from '../enums/newtroks.status.enum';
+import { NetworkStatus } from '../enums/network-status.enum';
 
 export class NotFoundError extends BaseApiError {
   constructor(message: string, code: ApiErrorCode) {
