+++ src/errors/enums/network-status.enum.ts
